add some features, need to test

+++ shop.go
@@ -3,7 +3,7 @@ package main
 // AutoCracker struct for the primary source of automatically generating value
 type AutoCracker struct {
 	name  string
-	price int
+	price float64
 	rate  int
 	held  int
 }
